+++ src/Extension.php
@@ -5,19 +5,11 @@
  * Description:
  * Copyright: Copyright (c) 2005 - 2015
  * Company: Pronamic
- * @author Stefan Boonstra
- * @version 1.0.0
+ * @author Remco Tolsma
+ * @version 1.1.0
+ * @since 1.0.0
  */
 class Pronamic_WP_Pay_Extensions_EDD_Extension {
-	/**
-	 * Slug
-	 *
-	 * @var string
-	 */
-	const SLUG = 'pronamic_ideal';
-
-	//////////////////////////////////////////////////
-
 	/**
 	 * Bootstrap
 	 */
@@ -33,166 +25,26 @@ class Pronamic_WP_Pay_Extensions_EDD_Extension {
 	 */
 	public static function plugins_loaded() {
 		if ( Pronamic_WP_Pay_Extensions_EDD_EasyDigitalDownloads::is_active() ) {
-			// Pronamic iDEAL Remove CC Form
-			// @see https://github.com/easydigitaldownloads/Easy-Digital-Downloads/blob/1.9.4/includes/checkout/template.php#L97
-			// @see https://github.com/easydigitaldownloads/Easy-Digital-Downloads/blob/1.9.4/includes/gateways/paypal-standard.php#L12
-			add_action( 'edd_' . self::SLUG . '_cc_form', array( __CLASS__, 'payment_fields' ) );
+			// Gateways
+			// @since 1.1.0
+			new Pronamic_WP_Pay_Extensions_EDD_Gateway( array(
+				'id'             => 'pronamic_ideal',
+				'admin_label'    => __( 'Pronamic', 'pronamic_ideal' ),
+				'checkout_label' => __( 'iDEAL', 'pronamic_ideal' ),
+			) );
 
-			add_action( 'edd_gateway_' . self::SLUG, array( __CLASS__, 'process_purchase' ) );
+			new Pronamic_WP_Pay_Extensions_EDD_CreditCardGateway();
+			new Pronamic_WP_Pay_Extensions_EDD_DirectDebitGateway();
+			new Pronamic_WP_Pay_Extensions_EDD_IDealGateway();
+			new Pronamic_WP_Pay_Extensions_EDD_MiniTixGateway();
+			new Pronamic_WP_Pay_Extensions_EDD_MisterCashGateway();
+			new Pronamic_WP_Pay_Extensions_EDD_SofortGateway();
 
 			add_action( 'pronamic_payment_status_update_easydigitaldownloads', array( __CLASS__, 'status_update' ), 10, 2 );
 			add_filter( 'pronamic_payment_source_text_easydigitaldownloads', array( __CLASS__, 'source_text' ), 10, 2 );
 
-			// Filters
-			add_filter( 'edd_settings_gateways', array( __CLASS__, 'settings_gateways' ) );
-			add_filter( 'edd_payment_gateways' , array( __CLASS__, 'payment_gateways' ) );
-
 			// Icons
 			add_filter( 'edd_accepted_payment_icons', array( __CLASS__, 'accepted_payment_icons' ) );
-		}
-	}
-
-	//////////////////////////////////////////////////
-
-	/**
-	 * Add the gateway to Easy Digital Downloads
-	 *
-	 * @param mixed $gateways
-	 *
-	 * @return mixed $gateways
-	 */
-	public static function payment_gateways( $gateways ) {
-		$gateways[ self::SLUG ] = array(
-			'admin_label'    => __( 'Pronamic iDEAL', 'pronamic_ideal' ),
-			'checkout_label' => __( 'iDEAL', 'pronamic_ideal' ),
-			'supports'       => array( 'buy_now' ),
-		);
-
-		return $gateways;
-	}
-
-	//////////////////////////////////////////////////
-
-	/**
-	 * Add the iDEAL configuration settings to the Easy Digital Downloads payment gateways settings page.
-	 *
-	 * @param mixed $settings_gateways
-	 *
-	 * @return mixed $settings_gateways
-	 */
-	public static function settings_gateways( $settings_gateways ) {
-		$settings_gateways[ self::SLUG ] = array(
-			'id'   => self::SLUG,
-			'name' => '<strong>' . __( 'iDEAL Settings', 'pronamic_ideal' ) . '</strong>',
-			'desc' => __( 'Configure the iDEAL settings', 'pronamic_ideal' ),
-			'type' => 'header',
-		);
-
-		$settings_gateways[ self::SLUG . '_config_id' ] = array(
-			'id'      => self::SLUG . '_config_id',
-			'name'    => __( 'iDEAL Configuration', 'pronamic_ideal' ),
-			'type'    => 'select',
-			'options' => Pronamic_WP_Pay_Plugin::get_config_select_options(),
-		);
-
-		return $settings_gateways;
-	}
-
-	//////////////////////////////////////////////////
-
-	/**
-	 * Payment fields for this gateway
-	 *
-	 * @see https://github.com/easydigitaldownloads/Easy-Digital-Downloads/blob/1.9.4/includes/checkout/template.php#L167
-	 */
-	public static function payment_fields() {
-		$gateway = Pronamic_WP_Pay_Plugin::get_gateway( edd_get_option( self::SLUG . '_config_id' ) );
-
-		if ( $gateway ) {
-			$input = $gateway->get_input_html();
-
-			if ( $input ) {
-				echo '<fieldset id="edd_cc_fields" class="edd-do-validate">';
-				echo '<span><legend>', __( 'iDEAL', 'pronamic_ideal' ), '</legend></span>';
-				echo $input;
-				echo '</fieldset>';
-			}
-		}
-	}
-
-	//////////////////////////////////////////////////
-
-	/**
-	 * The $purchase_data array consists of the following data:
-	 *
-	 * $purchase_data = array(
-	 *   'downloads'    => array of download IDs,
-	 *   'tax'          => taxed amount on shopping cart
-	 *   'subtotal'     => total price before tax
-	 *   'price'        => total price of cart contents after taxes,
-	 *   'purchase_key' => Random key
-	 *   'user_email'   => $user_email,
-	 *   'date'         => date( 'Y-m-d H:i:s' ),
-	 *   'user_id'      => $user_id,
-	 *   'post_data'    => $_POST,
-	 *   'user_info'    => array of user's information and used discount code
-	 *   'cart_details' => array of cart details,
-	 * );
-	 */
-	public static function process_purchase( $purchase_data ) {
-		$config_id = edd_get_option( self::SLUG . '_config_id' );
-
-		// Collect payment data
-		$payment_data = array(
-			'price'        => $purchase_data['price'],
-			'date'         => $purchase_data['date'],
-			'user_email'   => $purchase_data['user_email'],
-			'purchase_key' => $purchase_data['purchase_key'],
-			'currency'     => edd_get_currency(),
-			'downloads'    => $purchase_data['downloads'],
-			'user_info'    => $purchase_data['user_info'],
-			'cart_details' => $purchase_data['cart_details'],
-			'gateway'      => self::SLUG,
-			'status'       => 'pending',
-		);
-
-		// Record the pending payment
-		$payment_id = edd_insert_payment( $payment_data );
-
-		// Check payment
-		if ( ! $payment_id ) {
-			// Log error
-			edd_record_gateway_error( __( 'Payment Error', 'pronamic_ideal' ), sprintf( __( 'Payment creation failed before sending buyer to the iDEAL provider. Payment data: %s', 'pronamic_ideal' ), json_encode( $payment_data ) ), $payment_id );
-
-			edd_send_back_to_checkout( '?payment-mode=' . $purchase_data['post_data']['edd-gateway'] );
-		} else {
-			$data = new Pronamic_WP_Pay_Extensions_EDD_PaymentData( $payment_id, $payment_data );
-
-			$gateway = Pronamic_WP_Pay_Plugin::get_gateway( $config_id );
-
-			if ( $gateway ) {
-				// Start
-				$payment = Pronamic_WP_Pay_Plugin::start( $config_id, $gateway, $data );
-
-				$error = $gateway->get_error();
-
-				if ( is_wp_error( $error ) ) {
-					edd_record_gateway_error( __( 'Payment Error', 'pronamic_ideal' ), sprintf( __( 'Payment creation failed before sending buyer to the iDEAL provider. Payment data: %s', 'pronamic_ideal' ), json_encode( $payment_data ) ), $payment_id );
-
-					edd_set_error( 'pronamic_pay_error', Pronamic_WP_Pay_Plugin::get_default_error_message() );
-
-					foreach ( $error->get_error_messages() as $i => $message ) {
-						edd_set_error( 'pronamic_pay_error_' . $i, $message );
-					}
-
-					edd_send_back_to_checkout( '?payment-mode=' . $purchase_data['post_data']['edd-gateway'] );
-				} else {
-					// Redirect
-					$gateway->redirect( $payment );
-
-					exit;
-				}
-			}
 		}
 	}
 
@@ -242,7 +94,14 @@ class Pronamic_WP_Pay_Extensions_EDD_Extension {
 				break;
 			case Pronamic_WP_Pay_Statuses::SUCCESS :
 				if ( $should_update ) {
-					edd_insert_payment_note( $source_id, __( 'iDEAL payment completed.', 'pronamic_ideal' ) );
+					edd_insert_payment_note( $source_id, __( 'Payment completed.', 'pronamic_ideal' ) );
+
+					/*
+					 * @see https://github.com/easydigitaldownloads/Easy-Digital-Downloads/blob/2.2.8/includes/admin/payments/view-order-details.php#L36
+					 * @see https://github.com/easydigitaldownloads/Easy-Digital-Downloads/blob/2.2.8/includes/admin/payments/view-order-details.php#L199-L206
+					 * @see https://github.com/easydigitaldownloads/Easy-Digital-Downloads/blob/2.2.8/includes/payments/functions.php#L1312-L1332
+					 * @see https://github.com/easydigitaldownloads/Easy-Digital-Downloads/blob/2.2.8/includes/gateways/paypal-standard.php#L555-L576
+					 */
 				}
 
 				edd_update_payment_status( $source_id, Pronamic_WP_Pay_Extensions_EDD_EasyDigitalDownloads::ORDER_STATUS_PUBLISH );
@@ -254,13 +113,13 @@ class Pronamic_WP_Pay_Extensions_EDD_Extension {
 				break;
 			case Pronamic_WP_Pay_Statuses::OPEN :
 				if ( $should_update ) {
-					edd_insert_payment_note( $source_id, __( 'iDEAL payment open.', 'pronamic_ideal' ) );
+					edd_insert_payment_note( $source_id, __( 'Payment open.', 'pronamic_ideal' ) );
 				}
 
 				break;
 			default:
 				if ( $should_update ) {
-					edd_insert_payment_note( $source_id, __( 'iDEAL payment unknown.', 'pronamic_ideal' ) );
+					edd_insert_payment_note( $source_id, __( 'Payment unknown.', 'pronamic_ideal' ) );
 				}
 
 				break;