+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.1.0
+*	Added Credit Card gateway.
+*	Added Direct Debit gateway.
+*	Added iDEAL gateway.
+*	Added MiniTix gateway.
+*	Added Bancontact/Mister Cash gateway.
+*	Added SOFORT Banking gateway.
+*	Added gateway setting for the checkout label.
+*	Only show transaction ID if set.
+*	Added pending payment note with link to payment post.
+*	Tested on Easy Digital Downloads version 2.3.
+
 ## 1.0.1
 *	Changed WordPress pay core library requirment from ~1.0.0 to >=1.0.0.